Move loaders to its own file
