+++ HackrNews/Hacker News Feature/LiveHackrNewsLoader.swift
@@ -1,0 +1,10 @@
+//
+//  Copyright © 2020 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+public protocol LiveHackrNewsLoader {
+    typealias Result = Swift.Result<[LiveHackrNew], Swift.Error>
+    func load(completion: @escaping (Result) -> Void)
+}
